Move to shared and support more link
